requirements.txt
Flask==2.3.1
pymongo==4.3.3
waitress==2.1.2
requests
python-dotenv